requirements.txt
python_telegram_bot == 7.0.1
